moved methods into authentifcation
