+++ site-php/src/handlers/changeEmail.php
@@ -3,7 +3,6 @@
 namespace WebtoonLike\Site\handler;
 
 use WebtoonLike\Site\core\Authentication;
-use WebtoonLike\Site\core\editAccount;
 use WebtoonLike\Site\core\Router;
 
 
@@ -11,7 +10,7 @@ if (!isset($_POST['email']) || empty($_POST['email'])) {
     Router::redirect('/user', 301, ['error' => 'Le champ est vide.']);
 }
 
-$res = editAccount::editEmail($_SESSION['id'], $_POST['email']);
+$res = Authentication::editEmail($_SESSION['id'], $_POST['email']);
 
 if($res === true) {
     header("Location: /user");
