+++ site-php/src/handlers/changeUsername.php
@@ -2,7 +2,7 @@
 
 namespace WebtoonLike\Site\handler;
 
-use WebtoonLike\Site\core\editAccount;
+use WebtoonLike\Site\core\Authentication;
 use WebtoonLike\Site\core\Router;
 
 
@@ -10,7 +10,7 @@ if (!isset($_POST['username']) || empty($_POST['username'])) {
     Router::redirect('/user', 301, ['error' => 'Le champ est vide.']);
 }
 
-$res = editAccount::editUsername($_SESSION['id'], $_POST['username']);
+$res = Authentication::editUsername($_SESSION['id'], $_POST['username']);
 
 if($res === true) {
     header("Location: /user");
