+++ site-php/src/handlers/deleteAccount.php
@@ -3,7 +3,6 @@
 namespace WebtoonLike\Site\handler;
 
 use WebtoonLike\Site\core\Authentication;
-use WebtoonLike\Site\core\deleteAccount;
 use WebtoonLike\Site\core\Router;
 
 if ( !isset($_POST['deleteAccount'], $_POST['password'], $_SESSION['id'])
@@ -13,7 +12,7 @@ if ( !isset($_POST['deleteAccount'], $_POST['password'], $_SESSION['id'])
     Router::redirect('/deleteAccount', 301, ['errorChampVide' => 'Veuillez saisir les champs']);
 }
 
-$res = deleteAccount::deleteAccount($_SESSION['id'], $_POST['password']);
+$res = Authentication::deleteAccount($_SESSION['id'], $_POST['password']);
 
 if($res) {
     Authentication::logout();
